tweak: code style

+++ contracts/modules/IV1FYDai.sol
@@ -4,6 +4,6 @@ pragma solidity 0.8.6;
 import "@yield-protocol/utils-v2/contracts/token/IERC20.sol";
 
 interface IV1FYDai is IERC20 {
-    function maturity() external view returns(uint);
+    function maturity() external view returns(uint256);
     function redeem(address, address, uint256) external returns (uint256);
 }
